Rearanging file structure for Api project
